actualizacion modulo

+++ pet/src/main/java/com/pethistory/pet/models/Rol.java
@@ -1,6 +1,6 @@
 package com.pethistory.pet.models;
 
-import jakarta.persistence.Column;
+
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
@@ -12,6 +12,6 @@ import lombok.Getter;
 public class Rol {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer idRol;
+    private Long idRol;
     private String nomRol;
 }
